backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
neo4j==5.15.0
python-dotenv==1.0.1
pydantic==2.6.1
xmltodict==0.13.0
python-multipart==0.0.9
tqdm==4.66.2 